kernel/syscall: Rename handlers.rs to class0.rs
The syscall handlers in handlers.rs currently is for class0 so rename
the file to class0.rs. The plan is to introduce other syscall handlers
files for different classes.
